Add overall package documentation
Adds examples for the server and dequeuer at the top level. This made it clear
that the JobProcessor should have a function that provides it with a reasonable
default for the timeout, instead of 0. Adds NewJobProcessor to do that, and
updates some tests and other code to use it.

Fixes more package-level documentation.