Log unknown errors
In this case I was hitting this error because I was returning the empty string
for a HTTP response, instead of returning a decodable JSON object and it was
difficult to track down why this was happening. Let's log an error message if
we hit an unknown error so we can see why we are hitting this error case.

Also fixes the docstring for queued_jobs.Get.