services: better handling of timeouts
We can detect in some cases when a request timeout failed before the
request reached the downstream server, so do so. Also make a test fail
in a predictable way instead of an unpredictable way; see comment
inline for more info.

Remove a test that we can't stop from flaking.